Lazy-load proxy + instrumentation hot-path: 159.49 KB / 50.67 KB gzip
Two more lazy-load splits on top of the previous round. Hello-world goes
from 160.11 KB / 50.88 KB → 159.49 KB / 50.67 KB. The classic
v2-app-template also gets a small win (319.55 → 318.27 KB raw,
99.38 → 98.94 KB gzip).

Cumulative from baseline: 243.30 KB / 77.32 KB → 159.49 KB / 50.67 KB
(34.5% gzip reduction).

1. **`contentFor` extracted to `runtime/lib/mixins/content-for.ts`.**
   `each-in.ts` (which the renderer registers as the `-each-in` keyword
   helper) imports `contentFor` to unwrap proxies before iterating.
   Until now that import dragged in `runtime/lib/mixins/-proxy`, which
   defines `ProxyMixin = Mixin.create(...)` at module scope — the entry
   point to the entire EmberObject / Mixin / computed graph (proxy.ts
   imports `Mixin`, `computed`, `defineProperty`, `set`, etc.). Moving
   the 8-line `contentFor` function into its own file lets the renderer
   path keep proxy support without paying for the rest of the proxy
   mixin's transitive imports. `-proxy.ts` re-exports `contentFor` from
   the new file for back-compat.

2. **`@ember/instrumentation` hot path extracted to
   `instrumentation/lib/internal-instrument.ts`.** `_instrumentStart`
   (called by the resolver, curly manager, outlet/root/route-template
   managers) and `flaggedInstrument` (called by the views state
   machine) used to live in `index.ts` alongside `subscribe`,
   `unsubscribe`, `instrument`, etc. — most of which are dead code
   unless something actually subscribes (e.g. Ember Inspector). Moved
   `_instrumentStart`, `flaggedInstrument`, `subscribers`, the cache
   helpers, and `NOOP` to the lib file, with `index.ts` re-exporting
   them via `export ... from`. The `no-barrel-imports` autofix then
   rewrites internal callers to deep-import from the lib file. Net
   result: the `instrumentation/index.js` chunk (subscribe / unsubscribe
   / instrument machinery) drops out of bundles that only use the hot
   path.

`package.json`'s `ember-addon.renamed-modules` map gains an entry for
`runtime/lib/mixins/content-for.js` — that's emitted automatically by
the `packageMeta` rollup plugin, no manual edit.

Verified: lint clean, all type-checks pass, `test:node` 20/20,
`test:blueprints` 265/265, both smoke-test apps build.
